refactor: improved logging, error handling

+++ gulp-tasks/gulp-todo.js
@@ -7,14 +7,20 @@ const through2 = require('through2');
 const todo = require('gulp-todo');
 
 /**
+ * @function buildTodo
  * @description A function to generate a TODO.md file
  * @param {object} params - The function parameters.
  * @returns {Stream} Compiled file
  */
+
 const buildTodo = (params = {}) => {
   let todoExist = false;
-
   const filePath = './TODO.md';
+  const cb = params.cb || (() => {});
+
+  if (typeof cb !== 'function') {
+    throw new Error('Callback in params should be of type function.');
+  }
 
   if (fs.existsSync(filePath)) {
     fs.unlinkSync(filePath);
@@ -26,19 +32,19 @@ const buildTodo = (params = {}) => {
     .pipe(todo())
     .pipe(replace('### ', '# '))
     .pipe(
-      through2.obj(function processFile(file, _, cb) {
+      through2.obj(function processFile(file, _, callback) {
         if (file.todos?.length) {
           todoExist = true;
           this.push(file);
         }
-        cb();
+        callback();
       })
     )
     .pipe(gulp.dest('./'))
     .on('end', async () => {
       if (!todoExist) {
         log('         No TODOs found.');
-        params.cb?.();
+        cb();
         return;
       }
 
@@ -52,7 +58,7 @@ const buildTodo = (params = {}) => {
       } catch (error) {
         console.error(`exec error: ${error}`);
       }
-      params.cb?.();
+      cb();
     });
 };
